begins refactoring for flow
